requirements.txt
tensorflow==2.16.1
streamlit==1.12.0
numpy==1.26.4
matplotlib==3.5.3
seaborn==0.11.2
scikit-learn==1.0.2
pillow==9.2.0